fix: simplify nav bar and load chat on index route
