+++ js-peer/src/components/nav.tsx
@@ -1,18 +1,12 @@
 import { Fragment } from 'react'
-import { Disclosure, Menu, Transition } from '@headlessui/react'
+import { Disclosure, DisclosureButton, DisclosurePanel, Menu, Transition } from '@headlessui/react'
 import { Bars3Icon, BellIcon, XMarkIcon } from '@heroicons/react/24/outline'
 import Link from 'next/link'
 import Image from 'next/image'
 import { useRouter } from 'next/router'
 
 const navigationItems = [
-  { name: 'Chat', href: '/chat' },
   { name: 'Source', href: 'https://github.com/libp2p/universal-connectivity' },
-]
-const userNavigation = [
-  { name: 'Your Profile', href: '#' },
-  { name: 'Settings', href: '#' },
-  { name: 'Sign out', href: '#' },
 ]
 
 function classNames(...classes: string[]) {
@@ -27,8 +21,8 @@ export default function Navigation({ connectionInfoButton }: { connectionInfoBut
       {({ open }) => (
         <>
           <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
-            <div className="flex h-16 justify-between">
-              <div className="flex">
+            <div className="flex h-16 justify-between items-center">
+              <div className="flex items-center">
                 <div className="flex flex-shrink-0 items-center">
                   <Image src="/libp2p-logo.svg" alt="libp2p logo" height="46" width="46" />
                   <div className="ml-3 flex items-center">
@@ -36,7 +30,10 @@ export default function Navigation({ connectionInfoButton }: { connectionInfoBut
                     <Image src="/libp2p-hero.svg" alt="libp2p hero" height="24" width="24" className="ml-2 hidden sm:block" />
                   </div>
                 </div>
-                <div className="hidden sm:-my-px sm:ml-6 sm:flex sm:space-x-8">
+              </div>
+              
+              <div className="flex items-center space-x-4">
+                <div className="flex space-x-4">
                   {navigationItems.map((item) => (
                     <Link key={item.href} href={item.href} legacyBehavior>
                       <a
@@ -45,7 +42,7 @@ export default function Navigation({ connectionInfoButton }: { connectionInfoBut
                           router.pathname === item.href
                             ? 'border-indigo-500 text-gray-900'
                             : 'border-transparent text-gray-500 hover:border-gray-300 hover:text-gray-700',
-                          'inline-flex items-center border-b-2 px-1 pt-1 text-sm font-medium',
+                          'inline-flex items-center px-1 pt-1 text-sm font-medium',
                         )}
                         aria-current={router.pathname === item.href ? 'page' : undefined}
                       >
@@ -54,100 +51,12 @@ export default function Navigation({ connectionInfoButton }: { connectionInfoBut
                     </Link>
                   ))}
                 </div>
-              </div>
-              <div className="hidden sm:ml-6 sm:flex sm:items-center">
-                {connectionInfoButton}
-                
-                {/* <button
-                  type="button"
-                  className="ml-3 rounded-full bg-white p-1 text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                >
-                  <span className="sr-only">View notifications</span>
-                  <BellIcon className="h-6 w-6" aria-hidden="true" />
-                </button> */}
-
-                {/* Profile dropdown */}
-                <Menu as="div" className="relative ml-3">
-                  <div>
-                  </div>
-                  <Transition
-                    as={Fragment}
-                    enter="transition ease-out duration-200"
-                    enterFrom="transform opacity-0 scale-95"
-                    enterTo="transform opacity-100 scale-100"
-                    leave="transition ease-in duration-75"
-                    leaveFrom="transform opacity-100 scale-100"
-                    leaveTo="transform opacity-0 scale-95"
-                  >
-                    <Menu.Items className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
-                      {userNavigation.map((item) => (
-                        <Menu.Item key={item.name}>
-                          {({ active }) => (
-                            <a
-                              href={item.href}
-                              className={classNames(
-                                active ? 'bg-gray-100' : '',
-                                'block px-4 py-2 text-sm text-gray-700',
-                              )}
-                            >
-                              {item.name}
-                            </a>
-                          )}
-                        </Menu.Item>
-                      ))}
-                    </Menu.Items>
-                  </Transition>
-                </Menu>
-              </div>
-              <div className="-mr-2 flex items-center sm:hidden">
-                {/* Mobile menu button */}
-                <Disclosure.Button className="inline-flex items-center justify-center rounded-md bg-white p-2 text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2">
-                  <span className="sr-only">Open main menu</span>
-                  {open ? (
-                    <XMarkIcon className="block h-6 w-6" aria-hidden="true" />
-                  ) : (
-                    <Bars3Icon className="block h-6 w-6" aria-hidden="true" />
-                  )}
-                </Disclosure.Button>
+                <div className="flex items-center">
+                  {connectionInfoButton}
+                </div>
               </div>
             </div>
           </div>
-
-          <Disclosure.Panel className="sm:hidden">
-            <div className="space-y-1 pt-2 pb-3">
-              <div className="px-4 py-2">
-                <div className="flex items-center">
-                  <h1 className="text-lg font-semibold text-gray-900">Universal Connectivity</h1>
-                  <Image src="/libp2p-hero.svg" alt="libp2p hero" height="20" width="20" className="ml-2" />
-                </div>
-              </div>
-              {navigationItems.map((item) => (
-                <Link key={item.href} href={item.href} legacyBehavior>
-                  <Disclosure.Button
-                    key={item.href}
-                    className={classNames(
-                      router.pathname === item.href
-                        ? 'border-indigo-500 bg-indigo-50 text-indigo-700'
-                        : 'border-transparent text-gray-600 hover:border-gray-300 hover:bg-gray-50 hover:text-gray-800',
-                      'block border-l-4 py-2 pl-3 pr-4 text-base font-medium',
-                    )}
-                    aria-current={router.pathname === item.href ? 'page' : undefined}
-                  >
-                    {item.name}
-                  </Disclosure.Button>
-                </Link>
-              ))}
-            </div>
-            <div className="border-t border-gray-200 pt-4 pb-3">
-              <div className="flex items-center px-4">
-                {connectionInfoButton && (
-                  <div className="flex-shrink-0">
-                    {connectionInfoButton}
-                  </div>
-                )}
-              </div>
-            </div>
-          </Disclosure.Panel>
         </>
       )}
     </Disclosure>
